requirements.txt
# Core dependencies
flask>=3.0.0
python-dotenv>=1.0.0

# AI APIs
anthropic>=0.40.0
google-generativeai>=0.8.0
replicate>=0.25.0
openai>=1.30.0

# HTTP requests (for image downloads and direct API calls)
requests>=2.31.0

# Image processing
pillow>=10.0.0
